feat: Implement PM10 alert/advisory saving method
- 단순히 pm10 150, 300 이상만 저장
- 2시간 이상 지속 로직 추가 구현 필요

+++ fine-dust/src/main/java/com/example/finedust/repository/WarningRepository.java
@@ -3,5 +3,5 @@ package com.example.finedust.repository;
 import com.example.finedust.entity.WarningEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarningRepository extends JpaRepository<WarningEntity, Long> {
+public interface WarningRepository extends JpaRepository<WarningEntity, Integer> {
 }
